move useEagerConnect to libs/wallet, add logic for connecting state so token app can load after connection has succeeded or failed
