+++ apps/token/src/app-loader.tsx
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry/react';
 import { Splash } from '@vegaprotocol/ui-toolkit';
-import { useVegaWallet } from '@vegaprotocol/wallet';
+import { useVegaWallet, useEagerConnect } from '@vegaprotocol/wallet';
 import { useWeb3React } from '@web3-react/core';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
@@ -15,6 +15,7 @@ import {
 import { useContracts } from './contexts/contracts/contracts-context';
 import { useRefreshAssociatedBalances } from './hooks/use-refresh-associated-balances';
 import { getDataNodeUrl } from './lib/get-data-node-url';
+import { Connectors } from './lib/vega-connectors';
 
 export const AppLoader = ({ children }: { children: React.ReactElement }) => {
   const { t } = useTranslation();
@@ -24,14 +25,9 @@ export const AppLoader = ({ children }: { children: React.ReactElement }) => {
   const { token, staking, vesting } = useContracts();
   const setAssociatedBalances = useRefreshAssociatedBalances();
   const [balancesLoaded, setBalancesLoaded] = React.useState(false);
-  // TODO: TFE import detect when vega keys attempt complete
-  // const [vegaKeysLoaded, setVegaKeysLoaded] = React.useState(false);
+  const vegaConnecting = useEagerConnect(Connectors);
 
-  // Derive loaded state from all things that we want to load or attempted
-  // to load before rendering the app
-
-  // TODO: TFE import
-  const loaded = balancesLoaded; // && vegaKeysLoaded;
+  const loaded = balancesLoaded && !vegaConnecting;
 
   React.useEffect(() => {
     const run = async () => {
